+++ packages/create-lila-app/files/lila-universal.js
@@ -1,0 +1,10 @@
+/* eslint-disable import/no-unresolved, import/no-extraneous-dependencies */
+import tasksPlugin from 'lila-tasks';
+
+export default lila => {
+  tasksPlugin(lila);
+
+  return () => ({
+    tasks: [],
+  });
+};