+++ packages/lila-webpack-config/example/app/lila.js
@@ -1,0 +1,28 @@
+import tasksPlugin from '../../../lila-tasks/lib';
+import webpackPlugin from '../../../lila-webpack/lib';
+import webpackConfigPlugin from '../../lib';
+
+export default lila => {
+  tasksPlugin(lila);
+  webpackPlugin(lila);
+  webpackConfigPlugin(lila);
+
+  return ({ entry }) => ({
+    // devMiddleware: { writeToDisk: !0, watchOptions: { ignored: /node_modules/ } },
+    tasks: [
+      '@lila/webpack',
+      [
+        '@lila/move',
+        { source: 'build/index.html', target: `build/${entry}.html` },
+      ],
+    ],
+    alias: {
+      base: 'alias/base',
+      common: 'alias/common',
+    },
+    define: {
+      __DEFINE__: JSON.stringify('define'),
+    },
+    staticServer: 'https://www.google.com/lila',
+  });
+};