+++ rollup.config.js
@@ -9,7 +9,9 @@ const { name } = argv;
 const names =
   name === 'all'
     ? readdirSync('packages')
-    : name.split(',').map(n => `lila-${n}`);
+    : name
+        .split(',')
+        .map(n => (n === 'create' ? 'create-lila-app' : `lila-${n}`));
 
 export default names.map(n => ({
   input: `packages/${n}/src/index.js`,