+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.7.2 [2019-03-30]
+
+- Add `create-lila-app` plugin.
+- Add `lila-webpack-config` plugin.
+
 ## 0.7.1 [2019-03-11]
 
 - Add `lila-webpack` plugin.